Move document view check logic into reduce task
